requirements.txt
discord.py>=2.3
python-dotenv
google-generativeai
flask
google-genai
pillow